feat: Complete CSSTransitionManager implementation (#7462)
## Summary

This PR add a complete implementation of the `CSSTransitionManager`
class which will replace the current `CSSTransitionsRegistry` C++ class
and the current `CSSTransitionManager` JS class. It is much simpler and
more readable than the `CSSTransitionsRegistry`. The key difference is
that the `CSSTransitionManager` is responsible for management of just
one node's transition instead of managing all CSS transitions.