chore: Remove ShadowNode from CSSAnimation dependencies (#7470)
## Summary

This PR removes the shadow node instance stored in the `CSSAnimation`
props. This is needed as we won't store the shadow node in the animation
itself anymore. We are getting this instance when the node is cloned and
can simply pass it via props to calculate the current animation frame.