feat: Complete implementation of the native CSSAnimationsManger (#7473)
## Summary

This PR adds a complete implementation of the `CSSAnimationsManger`
class that is responsible for the view animations maangement (creating
new / removing old / updating changed / updating animation progress /
etc.).

It mainly does the same thing as the `CSSAnimationsRegistry` did before
but is way simpler and more readable. It will replace the
`CSSAnimationsRegistry` once the new implementation of CSS animations is
finished.